pyproject.toml
[project]
name = "serendipity"
version = "0.2.0"
description = "Personal Serendipity Engine - discover convergent and divergent content recommendations"
authors = [
    {name = "Warren Zhu", email = "wzhu@college.harvard.edu"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["cli", "discovery", "serendipity", "recommendations", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "claude-agent-sdk>=0.1.0",
    "pyperclip>=1.8.0",
    "aiohttp>=3.9.0",
    "questionary>=2.0.0",
    "rank-bm25>=0.2.0",
    "httpx>=0.25.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
serendipity = "serendipity.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["serendipity"]

[tool.hatch.build.targets.wheel.force-include]
"serendipity/templates" = "serendipity/templates"
"serendipity/prompts" = "serendipity/prompts"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["-v", "--strict-markers", "--tb=short"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]
